--- a/app/package.json
+++ b/app/package.json
@@ -22,6 +22,7 @@
     "git-describe": "3.0.2",
     "mkdirp": "0.5.1",
     "ms": "0.7.1",
+    "npm": "4.6.1",
     "node-fetch": "1.6.3",
     "node-pty": "0.6.6",
     "semver": "5.3.0",