--- a/app/plugins.js
+++ b/app/plugins.js
@@ -1,4 +1,3 @@
-const {exec} = require('child_process');
 const {resolve, basename} = require('path');
 const {writeFileSync} = require('fs');
 
@@ -6,7 +5,7 @@ const {app, dialog} = require('electron');
 const {sync: mkdirpSync} = require('mkdirp');
 const Config = require('electron-config');
 const ms = require('ms');
-const shellEnv = require('shell-env');
+const npm = require('npm');
 
 const config = require('./config');
 const notify = require('./notify');
@@ -224,46 +223,28 @@ function toDependencies(plugins) {
 }
 
 function install(fn) {
-  const {shell: cfgShell, npmRegistry} = exports.getDecoratedConfig();
+  const {npmRegistry} = exports.getDecoratedConfig();
+  const config = {
+    prefix: path,
+    production: true,
+    'scripts-prepend-node-path': false
+  };
 
-  const shell = cfgShell && cfgShell !== '' ? cfgShell : undefined;
+  if (npmRegistry) {
+    config.registry = npmRegistry;
+  }
 
-  shellEnv(shell).then(env => {
-    if (npmRegistry) {
-      env.NPM_CONFIG_REGISTRY = npmRegistry;
+  npm.load(config, err => {
+    if (err) {
+      return fn(err);
     }
-    /* eslint-disable camelcase  */
-    env.npm_config_runtime = 'electron';
-    env.npm_config_target = process.versions.electron;
-    env.npm_config_disturl = 'https://atom.io/download/atom-shell';
-    /* eslint-enable camelcase  */
-    // Shell-specific installation commands
-    const installCommands = {
-      fish: 'npm prune; and npm install --production --no-shrinkwrap',
-      posix: 'npm prune && npm install --production --no-shrinkwrap'
-    };
-    // determine the shell we're running in
-    const whichShell = (typeof cfgShell === 'string' && cfgShell.match(/fish/)) ? 'fish' : 'posix';
-    const execOptions = {
-      cwd: path,
-      env
-    };
-
-    // https://nodejs.org/api/child_process.html#child_process_child_process_exec_command_options_callback
-    // node.js requires command line parsing should be compatible with cmd.exe on Windows, should able to accept `/d /s /c`
-    // but most custom shell doesn't. Instead, falls back to default shell
-    if (process.platform !== 'win32') {
-      execOptions.shell = shell;
-    }
-
-    // Use the install command that is appropriate for our shell
-    exec(installCommands[whichShell], execOptions, err => {
-      if (err) {
-        return fn(err);
+    npm.commands.install([], er => {
+      if (er) {
+        return fn(er);
       }
-      fn(null);
+      return fn(null);
     });
-  }).catch(fn);
+  });
 }
 
 exports.subscribe = function (fn) {