With this PR, user won't need to have npm installed to have plugins.
Hyper will be protected against (npm@5) incompatibility too.

It works great but some tasks remain to be done :
* [x] Find a way to make a `npm prune`
* [x] Benchmark `Hyper`size and speed
* [ ] Test on Windows & Linux