@chabou 2017-06-02
Using `npm prune` programmatically is not trivial because there is no option to make it in another directory than cwd (like `prefix` option for `npm install` command).