@rauchg 2017-06-09
I really want this for the next release! Great work @chabou