@chabou 2017-06-09
Little benchmark. 

* Speed
Adding/removing 5 plugins when Hyper is started `['hyper-font-smoothing', 'hyper-simple-vibrancy', 'hyper-pane', 'hyper-broadcast', 'hyper-autoprofile']`
  * Master: adding 1693ms, removing 1362ms
  * This PR: adding 679ms, removing 473ms
* Exe size (on Windows)
  * Master: 64 396 288
  * This PR: 67 333 632

Speed improvements was not an objectif but it is pretty cool.
Size increase worth the system consistency

NPM v4.6.1 because v5.0.3 API is buggy. Callbacks are too many called.