@albinekb 2017-06-10
https://github.com/zeit/hyper/pull/172#issuecomment-239582144