@chabou 2017-06-10
Yes, using internal (public but non documented) API is bad. And I really understand @iarna disclaimer. 
But it will not be a problem to adapt our code to future changes. This PR is using a very little part of this API. It is simple and it works great.

However, I am totally ok if we prefer #381 approach.