@albinekb 2017-06-10
Yeah I think this is OK now, but we'll need to change it sooner or later, so better to go with the "final" solution right away (run it as cli)