@iarna 2017-06-10
You can get machine readable output from calling `npm prune --json` and `npm install --json`, albeit this is much improved with npm@5.