@chabou 2017-06-26
Closed in favor of #381 which is ready to be merged.